# src/openapi/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Taja Shop API",
    "description": "OpenAPI map of Next.js `/api/*` routes. Request/response bodies match Zod/Mongoose usage in each `route.ts`. Full business rules: **docs/BACKEND_BUSINESS_LOGIC.md**. Swagger UI: `/docs/api`.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Override via NEXT_PUBLIC_APP_URL"
    }
  ],
...
  "paths": {
...
    "/api/payments/verify": {
      "get": {
        "tags": [
          "Payments"
        ],
        "summary": "Verify transaction",
        "responses": {
          "200": {
            "description": "Success (see route implementation for shape)"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          }
        }
      },
      "post": {
        "tags": [
          "Payments"
        ],
        "summary": "Verify POST",
        "responses": {
          "200": {
            "description": "Webhook processed or already processed"
          },
          "400": {
            "description": "Reference missing"
          },
          "404": {
            "description": "Order not found"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PaymentVerifyWebhookRequest"
              }
            }
          }
        }
      }
    },
    "/api/payments/webhook/paystack": {
      "post": {
        "tags": [
          "Payments"
        ],
        "summary": "Paystack webhook",
        "responses": {
          "200": {
            "description": "Success (see route implementation for shape)"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Authorization: Bearer <token> where required"
      }
    },
    "schemas": {
...
      "PaymentVerifyWebhookRequest": {
        "type": "object",
        "properties": {
          "provider": {
            "type": "string",
            "enum": [
              "auto",
              "paystack",
              "flutterwave"
            ],
            "default": "auto"
          },
          "reference": {
            "type": "string"
          },
          "data": {
            "type": "object",
            "properties": {
              "reference": {
                "type": "string"
              },
              "tx_ref": {
                "type": "string"
              }
            }
          }
        }
      },
...
    }
  }
}